+++ examples/tool-helpers-json-schema.ts
@@ -6,7 +6,7 @@ import { betaTool } from 'openai/helpers/beta/json-schema';
 const client = new OpenAI();
 
 async function main() {
-  const message = await client.chat.completions.toolRunner({
+  const message = await client.beta.chat.completions.toolRunner({
     messages: [
       {
         role: 'user',
